requirements.txt
diffusers
flask
torch
transformers
accelerate
pillow
gradio==3.41.2
peft
